Extract "tagName" assertion to own file
